Refactored ComboBoxAndButton.java into ChannelSelectorPanel
* Removed testing and demo classes
